solve issues reported by code climate
